OpenAI API/requirements.txt
ipykernel==7.1.0
openai==2.7.1
python-dotenv==1.2.1
pypdf==6.1.3
pandas>=2.0.0
redis>=5.0.0
psycopg2-binary>=2.9.0

